Test_Project/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
python==3.7.6